pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robocomm"
version = "0.1.0"
keywords = ["comms", "robotics", "rl"]
maintainers = [
  { name="Soulde", email="fakesoulde@gmail.com" },
]
authors = [
  { name="Soulde", email="fakesoulde@gmail.com" },
]
description = "Simple communication framework for robotics based on zmq"
readme = { file = "README.md", content-type = "text/markdown"}
license = { text = "MIT-Clause" }

requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyzmq",
    "numpy"
]

[project.urls]
Homepage = "https://github.com/soulde/robocomm"
Issues = "https://github.com/soulde/robocomm/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["robocomm*"]

[tool.setuptools.package-data]


[tool.isort]

py_version = 37
line_length = 120
group_by_package = true

# Files to skip
skip_glob = [".vscode/*"]

# Order of imports
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

# Extra standard libraries considered as part of python (permissive licenses)
extra_standard_library = [
    "numpy",
    "pyzmq"

]
# Imports from this repository
known_first_party = "robocomm"

[tool.pyright]

include = ["robocomm"]

typeCheckingMode = "basic"
pythonVersion = "3.7"
pythonPlatform = "Linux"
enableTypeIgnoreComments = true

# This is required as the CI pre-commit does not download the module (i.e. numpy, torch, prettytable)
# Therefore, we have to ignore missing imports
reportMissingImports = "none"
# This is required to ignore for type checks of modules with stubs missing.
reportMissingModuleSource = "none" # -> most common: prettytable in mdp managers

reportGeneralTypeIssues = "none"       # -> raises 218 errors (usage of literal MISSING in dataclasses)
reportOptionalMemberAccess = "warning" # -> raises 8 errors
reportPrivateUsage = "warning"
